+++ src/Entity/PrdHistoEntrepot.php
@@ -25,7 +25,7 @@ class PrdHistoEntrepot
     private ?string $indice = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $stock = null;
+    private ?float $stock = null;
 
     #[ORM\Column(nullable: true)]
     private ?bool $statut = null;
@@ -96,12 +96,12 @@ class PrdHistoEntrepot
         return $this;
     }
 
-    public function getStock(): ?int
+    public function getStock(): ?float
     {
         return $this->stock;
     }
 
-    public function setStock(?int $stock): self
+    public function setStock(?float $stock): self
     {
         $this->stock = $stock;
